pimpl around all of main model
