+++ tests/CMakeLists.txt
@@ -6,7 +6,6 @@ find_package(doctest REQUIRED)
 
 include("${doctest_DIR}/doctest.cmake")
 
-add_subdirectory("power_grid_model_static")
 add_subdirectory("c_api_tests")
 add_subdirectory("cpp_unit_tests")
 add_subdirectory("cpp_integration_tests")
